'修复当返回类型为枚举集合时无法获取到TypeHandler'

+++ src/SmartSql.Test.Unit/DB/init-mysql-db.sql
@@ -16,7 +16,7 @@ create table T_AllPrimitive
     String                varchar(100) not null,
     Guid                  char(36)     not null,
     TimeSpan              time         not null,
-    NumericalEnum         tinyint(1)   not null,
+    NumericalEnum         smallint   not null,
     NullableBoolean       tinyint(1)   null,
     NullableChar          char         null,
     NullableInt16         mediumint    null,
@@ -27,7 +27,7 @@ create table T_AllPrimitive
     NullableDateTime      datetime     null,
     NullableGuid          char(36)     null,
     NullableTimeSpan      time         null,
-    NullableNumericalEnum tinyint(1)   null,
+    NullableNumericalEnum smallint   null,
     NullableString        varchar(100) null
 ) engine = InnoDb;
